LPS-67289 Uses form navigator entry to show the custom user attributes configuration on asset publisher
