+++ modules/apps/web-experience/asset/asset-entry-query-processor-custom-user-attributes/build.gradle
@@ -8,4 +8,5 @@ dependencies {
 	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
 	provided group: "javax.servlet.jsp", name: "javax.servlet.jsp-api", version: "2.3.1"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	provided project(":apps:web-experience:asset:asset-publisher-web")
 }
