+++ modules/apps/web-experience/asset/asset-entry-query-processor-custom-user-attributes/src/main/java/com/liferay/asset/entry/query/processor/custom/user/attributes/internal/CustomUserAttributesAssetEntryQueryProcessor.java
@@ -16,25 +16,23 @@ package com.liferay.asset.entry.query.processor.custom.user.attributes.internal;
 
 import com.liferay.asset.kernel.service.persistence.AssetEntryQuery;
 import com.liferay.asset.kernel.util.AssetEntryQueryProcessor;
-import com.liferay.asset.kernel.util.BaseJSPAssetEntryQueryProcessor;
 import com.liferay.asset.publisher.web.constants.AssetPublisherPortletKeys;
 import com.liferay.asset.publisher.web.util.AssetPublisherUtil;
-import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 
+import java.io.IOException;
+
 import java.util.Locale;
-import java.util.ResourceBundle;
 
 import javax.portlet.PortletPreferences;
 
-import javax.servlet.ServletContext;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Jorge Ferrer
@@ -47,19 +45,23 @@ import org.osgi.service.component.annotations.Reference;
 	service = AssetEntryQueryProcessor.class
 )
 public class CustomUserAttributesAssetEntryQueryProcessor
-	extends BaseJSPAssetEntryQueryProcessor {
+	implements AssetEntryQueryProcessor {
 
 	@Override
-	public String getJspPath() {
-		return "/custom_user_attributes.jsp";
+	public String getKey() {
+		return null;
 	}
 
 	@Override
 	public String getTitle(Locale locale) {
-		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
-			"content.Language", locale, getClass());
+		return null;
+	}
 
-		return LanguageUtil.get(resourceBundle, "custom-user-attributes");
+	@Override
+	public void include(
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse)
+		throws IOException {
 	}
 
 	@Override
@@ -73,15 +75,6 @@ public class CustomUserAttributesAssetEntryQueryProcessor
 
 		AssetPublisherUtil.addUserAttributes(
 			user, StringUtil.split(customUserAttributes), assetEntryQuery);
-	}
-
-	@Override
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.asset.entry.query.processor.custom.user.attributes)",
-		unbind = "-"
-	)
-	public void setServletContext(ServletContext servletContext) {
-		super.setServletContext(servletContext);
 	}
 
 }
